remove colormap from brat dataset
Fiji apparently doesn't save a full colormap along with the image. This causes the numpy/PIL approach used here to fail as it seems to truncate the resulting image, resulting in only black pixels. Thus enforce a 256 color palette, remap the ignored label to 255 and shift root and shoot labels by 1. In future, this should rather be dealt with in BRAT itself